+++ src/com/jgarms/adventOfCode2022/day11/OperationType.java
@@ -3,20 +3,20 @@ package com.jgarms.adventOfCode2022.day11;
 public enum OperationType {
     ADD() {
         @Override
-        public int performOperation(int old, Integer amount) {
+        public long performOperation(long old, Long amount) {
             return old + amount;
         }
     }, MULTIPLY {
         @Override
-        public int performOperation(int old, Integer amount) {
+        public long performOperation(long old, Long amount) {
             return old * amount;
         }
     }, SQUARE {
         @Override
-        public int performOperation(int old, Integer amount) {
+        public long performOperation(long old, Long amount) {
             return old * old;
         }
     };
 
-    public abstract int performOperation(int old, Integer amount);
+    public abstract long performOperation(long old, Long amount);
 }
